+++ crates/rust-client/src/transaction/request/builder.rs
@@ -1,5 +1,9 @@
 //! Contains structures and functions related to transaction creation.
-use alloc::{collections::BTreeMap, string::ToString, vec::Vec};
+use alloc::{
+    collections::{BTreeMap, BTreeSet},
+    string::ToString,
+    vec::Vec,
+};
 
 use miden_lib::note::{create_p2id_note, create_p2ide_note, create_swap_note};
 use miden_objects::{
@@ -32,7 +36,7 @@ pub struct TransactionRequestBuilder {
     unauthenticated_input_notes: Vec<Note>,
     /// Notes to be consumed by the transaction together with their (optional) arguments. This
     /// includes both authenticated and unauthenticated notes.
-    input_notes: BTreeMap<NoteId, Option<NoteArgs>>,
+    input_notes: Vec<(NoteId, Option<NoteArgs>)>,
     /// Notes to be created by the transaction. This includes both full and partial output notes.
     /// The transaction script will be generated based on these notes.
     own_output_notes: Vec<OutputNote>,
@@ -74,7 +78,7 @@ impl TransactionRequestBuilder {
     pub fn new() -> Self {
         Self {
             unauthenticated_input_notes: vec![],
-            input_notes: BTreeMap::new(),
+            input_notes: vec![],
             own_output_notes: Vec::new(),
             expected_output_recipients: BTreeMap::new(),
             expected_future_notes: BTreeMap::new(),
@@ -95,7 +99,7 @@ impl TransactionRequestBuilder {
         notes: impl IntoIterator<Item = (Note, Option<NoteArgs>)>,
     ) -> Self {
         for (note, argument) in notes {
-            self.input_notes.insert(note.id(), argument);
+            self.input_notes.push((note.id(), argument));
             self.unauthenticated_input_notes.push(note);
         }
         self
@@ -108,7 +112,7 @@ impl TransactionRequestBuilder {
         notes: impl IntoIterator<Item = (NoteId, Option<NoteArgs>)>,
     ) -> Self {
         for (note_id, argument) in notes {
-            self.input_notes.insert(note_id, argument);
+            self.input_notes.push((note_id, argument));
         }
         self
     }
@@ -357,6 +361,13 @@ impl TransactionRequestBuilder {
     /// - If an expiration delta is set when a custom script is set.
     /// - If an invalid note variant is encountered in the own output notes.
     pub fn build(self) -> Result<TransactionRequest, TransactionRequestError> {
+        let mut seen_input_notes = BTreeSet::new();
+        for (note_id, _) in &self.input_notes {
+            if !seen_input_notes.insert(note_id) {
+                return Err(TransactionRequestError::DuplicateInputNote(*note_id));
+            }
+        }
+
         let script_template = match (self.custom_script, self.own_output_notes.is_empty()) {
             (Some(_), false) => {
                 return Err(TransactionRequestError::ScriptTemplateError(